Access basic application statistics
Most of this change is vendoring a load of things.

But it also breaks out the usage of the CF CLI API. The UI stuff bleeds
through a bit, but works for now.

Need to tie down the dependencies in the manifest so that it’s more
apparent what is being used (and thus what should be upgraded as
security things are addressed).